+++ megalinter/linters/CSpellLinter.py
@@ -5,7 +5,6 @@ https://github.com/nvuillam/npm-groovy-lint
 """
 import json
 import logging
-import os.path
 import re
 
 from megalinter import Linter
@@ -49,4 +48,4 @@ Of course, please correct real typos before :)
         logging.debug(
             f"Generated additional TextReporter log for CSpellLinter:\n{additional_report}"
         )
-        return additional_report.split(os.linesep)
+        return additional_report.splitlines()